Refactor: add syncer server
