+++ etc/conf/microservice.yaml
@@ -1,4 +1,4 @@
 servicecomb:
   service:
-    name: servicecomb-service-center
-    version: 2.0.0
+    name: SERVICECENTER
+    version: 2.1.0
