+++ etc/conf/syncer.yaml
@@ -1,14 +1,8 @@
 sync:
-  enableOnStart: true
+  enableOnStart: false
   peers:
-    - name: dc2
+    - name: dc
       kind: servicecomb
       endpoints: ["127.0.0.1:30105"]
       # only allow mode implemented in incremental approach like push, watch(such as pub/sub, long polling)
       mode: [push]
-      revision: 100
-    - name: dc3
-      kind: consul
-      endpoints: []
-      # since consul will not push data to servcecomb, if we need set push and watch mode to achieve two direction sync
-      mode: [push,watch]
